requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
xgboost
python-pptx
jupyter
ipykernel
